[REFACTOR]내가속한 프로젝트 조회 응답 데이터 수정
